Flatten all files
