+++ agent/consul/gateways/controller_gateways_test.go
@@ -18,6 +18,1310 @@ import (
 	"github.com/stretchr/testify/require"
 )
 
+func TestBoundAPIGatewayBindRoute(t *testing.T) {
+	t.Parallel()
+
+	cases := map[string]struct {
+		gateway              gatewayMeta
+		route                structs.BoundRoute
+		expectedBoundGateway structs.BoundAPIGatewayConfigEntry
+		expectedDidBind      bool
+		expectedErr          error
+	}{
+		"Bind TCP Route to Gateway": {
+			gateway: gatewayMeta{
+				BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+					Kind: structs.BoundAPIGateway,
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener",
+							Routes: []structs.ResourceReference{},
+						},
+					},
+				},
+				Gateway: &structs.APIGatewayConfigEntry{
+					Kind: structs.APIGateway,
+					Name: "Gateway",
+					Listeners: []structs.APIGatewayListener{
+						{
+							Name:     "Listener",
+							Protocol: structs.ListenerProtocolTCP,
+						},
+					},
+				},
+			},
+			route: &structs.TCPRouteConfigEntry{
+				Kind: structs.TCPRoute,
+				Name: "Route",
+				Parents: []structs.ResourceReference{
+					{
+						Kind:        structs.APIGateway,
+						Name:        "Gateway",
+						SectionName: "Listener",
+					},
+				},
+			},
+			expectedBoundGateway: structs.BoundAPIGatewayConfigEntry{
+				Kind: structs.BoundAPIGateway,
+				Name: "Gateway",
+				Listeners: []structs.BoundAPIGatewayListener{
+					{
+						Name: "Listener",
+						Routes: []structs.ResourceReference{
+							{
+								Kind: structs.TCPRoute,
+								Name: "Route",
+							},
+						},
+					},
+				},
+			},
+			expectedDidBind: true,
+		},
+		"Bind TCP Route with wildcard section name to all listeners on Gateway": {
+			gateway: gatewayMeta{
+				BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+					Kind: structs.BoundAPIGateway,
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener 1",
+							Routes: []structs.ResourceReference{},
+						},
+						{
+							Name:   "Listener 2",
+							Routes: []structs.ResourceReference{},
+						},
+						{
+							Name:   "Listener 3",
+							Routes: []structs.ResourceReference{},
+						},
+					},
+				},
+				Gateway: &structs.APIGatewayConfigEntry{
+					Kind: structs.APIGateway,
+					Name: "Gateway",
+					Listeners: []structs.APIGatewayListener{
+						{
+							Name:     "Listener 1",
+							Protocol: structs.ListenerProtocolTCP,
+						},
+						{
+							Name:     "Listener 2",
+							Protocol: structs.ListenerProtocolTCP,
+						},
+						{
+							Name:     "Listener 3",
+							Protocol: structs.ListenerProtocolTCP,
+						},
+					},
+				},
+			},
+			route: &structs.TCPRouteConfigEntry{
+				Kind: structs.TCPRoute,
+				Name: "Route",
+				Parents: []structs.ResourceReference{
+					{
+						Kind: structs.APIGateway,
+						Name: "Gateway",
+					},
+				},
+			},
+			expectedBoundGateway: structs.BoundAPIGatewayConfigEntry{
+				Kind: structs.BoundAPIGateway,
+				Name: "Gateway",
+				Listeners: []structs.BoundAPIGatewayListener{
+					{
+						Name: "Listener 1",
+						Routes: []structs.ResourceReference{
+							{
+								Kind: structs.TCPRoute,
+								Name: "Route",
+							},
+						},
+					},
+					{
+						Name: "Listener 2",
+						Routes: []structs.ResourceReference{
+							{
+								Kind: structs.TCPRoute,
+								Name: "Route",
+							},
+						},
+					},
+					{
+						Name: "Listener 3",
+						Routes: []structs.ResourceReference{
+							{
+								Kind: structs.TCPRoute,
+								Name: "Route",
+							},
+						},
+					},
+				},
+			},
+			expectedDidBind: true,
+		},
+		"TCP Route cannot bind to Gateway because the parent reference kind is not APIGateway": {
+			gateway: gatewayMeta{
+				BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+					Kind:      structs.BoundAPIGateway,
+					Name:      "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{},
+				},
+				Gateway: &structs.APIGatewayConfigEntry{
+					Kind:      structs.APIGateway,
+					Name:      "Gateway",
+					Listeners: []structs.APIGatewayListener{},
+				},
+			},
+			route: &structs.TCPRouteConfigEntry{
+				Kind: structs.TCPRoute,
+				Name: "Route",
+				Parents: []structs.ResourceReference{
+					{
+						Name:        "Gateway",
+						SectionName: "Listener",
+					},
+				},
+			},
+			expectedBoundGateway: structs.BoundAPIGatewayConfigEntry{
+				Kind:      structs.TerminatingGateway,
+				Name:      "Gateway",
+				Listeners: []structs.BoundAPIGatewayListener{},
+			},
+			expectedDidBind: false,
+		},
+		"TCP Route cannot bind to Gateway because the parent reference name does not match": {
+			gateway: gatewayMeta{
+				BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+					Kind:      structs.BoundAPIGateway,
+					Name:      "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{},
+				},
+				Gateway: &structs.APIGatewayConfigEntry{
+					Kind:      structs.APIGateway,
+					Name:      "Gateway",
+					Listeners: []structs.APIGatewayListener{},
+				},
+			},
+			route: &structs.TCPRouteConfigEntry{
+				Kind: structs.TCPRoute,
+				Name: "Route",
+				Parents: []structs.ResourceReference{
+					{
+						Kind:        structs.APIGateway,
+						Name:        "Other Gateway",
+						SectionName: "Listener",
+					},
+				},
+			},
+			expectedBoundGateway: structs.BoundAPIGatewayConfigEntry{
+				Kind:      structs.BoundAPIGateway,
+				Name:      "Gateway",
+				Listeners: []structs.BoundAPIGatewayListener{},
+			},
+			expectedDidBind: false,
+		},
+		"TCP Route cannot bind to Gateway because it lacks listeners": {
+			gateway: gatewayMeta{
+				BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+					Kind:      structs.BoundAPIGateway,
+					Name:      "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{},
+				},
+				Gateway: &structs.APIGatewayConfigEntry{
+					Kind:      structs.APIGateway,
+					Name:      "Gateway",
+					Listeners: []structs.APIGatewayListener{},
+				},
+			},
+			route: &structs.TCPRouteConfigEntry{
+				Kind: structs.TCPRoute,
+				Name: "Route",
+				Parents: []structs.ResourceReference{
+					{
+						Kind:        structs.APIGateway,
+						Name:        "Gateway",
+						SectionName: "Listener",
+					},
+				},
+			},
+			expectedBoundGateway: structs.BoundAPIGatewayConfigEntry{
+				Kind:      structs.BoundAPIGateway,
+				Name:      "Gateway",
+				Listeners: []structs.BoundAPIGatewayListener{},
+			},
+			expectedDidBind: false,
+			expectedErr:     fmt.Errorf("route cannot bind because gateway has no listeners"),
+		},
+		"TCP Route cannot bind to Gateway because it has an invalid section name": {
+			gateway: gatewayMeta{
+				BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+					Kind: structs.BoundAPIGateway,
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener",
+							Routes: []structs.ResourceReference{},
+						},
+					},
+				},
+				Gateway: &structs.APIGatewayConfigEntry{
+					Kind: structs.APIGateway,
+					Name: "Gateway",
+					Listeners: []structs.APIGatewayListener{
+						{
+							Name:     "Listener",
+							Protocol: structs.ListenerProtocolTCP,
+						},
+					},
+				},
+			},
+			route: &structs.TCPRouteConfigEntry{
+				Kind: structs.TCPRoute,
+				Name: "Route",
+				Parents: []structs.ResourceReference{
+					{
+						Kind:        structs.APIGateway,
+						Name:        "Gateway",
+						SectionName: "Other Listener",
+					},
+				},
+			},
+			expectedBoundGateway: structs.BoundAPIGatewayConfigEntry{
+				Kind: structs.BoundAPIGateway,
+				Name: "Gateway",
+				Listeners: []structs.BoundAPIGatewayListener{
+					{
+						Name:   "Listener",
+						Routes: []structs.ResourceReference{},
+					},
+				},
+			},
+			expectedDidBind: false,
+			expectedErr:     fmt.Errorf("failed to bind route Route to gateway Gateway with listener 'Other Listener'"),
+		},
+	}
+
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			ref := tc.route.GetParents()[0]
+
+			actualDidBind, _, actualErrors := (&tc.gateway).initialize().updateRouteBinding(tc.route)
+
+			require.Equal(t, tc.expectedDidBind, actualDidBind)
+			require.Equal(t, tc.expectedErr, actualErrors[ref])
+			require.Equal(t, tc.expectedBoundGateway.Listeners, tc.gateway.BoundGateway.Listeners)
+		})
+	}
+}
+
+func TestBoundAPIGatewayUnbindRoute(t *testing.T) {
+	t.Parallel()
+
+	cases := map[string]struct {
+		gateway           gatewayMeta
+		route             structs.BoundRoute
+		expectedGateway   structs.BoundAPIGatewayConfigEntry
+		expectedDidUnbind bool
+	}{
+		"TCP Route unbinds from Gateway": {
+			gateway: gatewayMeta{
+				BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+					Kind: structs.BoundAPIGateway,
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener",
+							Routes: []structs.ResourceReference{
+								{
+									Kind: structs.TCPRoute,
+									Name: "Route",
+								},
+							},
+						},
+					},
+				},
+				Gateway: &structs.APIGatewayConfigEntry{},
+			},
+			route: &structs.TCPRouteConfigEntry{
+				Kind: structs.TCPRoute,
+				Name: "Route",
+				Parents: []structs.ResourceReference{
+					{
+						Kind:        structs.BoundAPIGateway,
+						Name:        "Gateway",
+						SectionName: "Listener",
+					},
+				},
+			},
+			expectedGateway: structs.BoundAPIGatewayConfigEntry{
+				Kind: structs.BoundAPIGateway,
+				Name: "Gateway",
+				Listeners: []structs.BoundAPIGatewayListener{
+					{
+						Name:   "Listener",
+						Routes: []structs.ResourceReference{},
+					},
+				},
+			},
+			expectedDidUnbind: true,
+		},
+	}
+
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			routeRef := structs.ResourceReference{
+				Kind:           tc.route.GetKind(),
+				Name:           tc.route.GetName(),
+				EnterpriseMeta: *tc.route.GetEnterpriseMeta(),
+			}
+			actualDidUnbind := (&tc.gateway).initialize().unbindRoute(routeRef)
+
+			require.Equal(t, tc.expectedDidUnbind, actualDidUnbind)
+			require.Equal(t, tc.expectedGateway.Listeners, tc.gateway.BoundGateway.Listeners)
+		})
+	}
+}
+
+func TestBindRoutesToGateways(t *testing.T) {
+	t.Parallel()
+
+	type testCase struct {
+		gateways                 []*gatewayMeta
+		routes                   []structs.BoundRoute
+		expectedBoundAPIGateways []*structs.BoundAPIGatewayConfigEntry
+		expectedReferenceErrors  map[structs.ResourceReference]error
+	}
+
+	cases := map[string]testCase{
+		"TCP Route binds to gateway": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name: "Listener",
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Kind: structs.TCPRoute,
+					Name: "TCP Route",
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route unbinds from gateway": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name: "Listener",
+								Routes: []structs.ResourceReference{
+									{
+										Name:        "TCP Route",
+										Kind:        structs.TCPRoute,
+										SectionName: "",
+									},
+								},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Kind:    structs.TCPRoute,
+					Name:    "TCP Route",
+					Parents: []structs.ResourceReference{},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener",
+							Routes: []structs.ResourceReference{},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route binds to multiple gateways": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway 1",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway 1",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway 2",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway 2",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Kind: structs.TCPRoute,
+					Name: "TCP Route",
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway 1",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener",
+						},
+						{
+							Name:        "Gateway 2",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway 1",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+				{
+					Name: "Gateway 2",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route binds to a single listener on a gateway with multiple listeners": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener 1",
+								Routes: []structs.ResourceReference{},
+							},
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolHTTP,
+							},
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener 2",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener 1",
+							Routes: []structs.ResourceReference{},
+						},
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route binds to all listeners on a gateway": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener 1",
+								Routes: []structs.ResourceReference{},
+							},
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener 1",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route binds to gateway with multiple listeners, one of which is already bound": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name: "Listener 1",
+								Routes: []structs.ResourceReference{
+									{
+										Name:        "TCP Route",
+										Kind:        structs.TCPRoute,
+										SectionName: "",
+									},
+								},
+							},
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener 1",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route binds to a listener on multiple gateways": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway 1",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener 1",
+								Routes: []structs.ResourceReference{},
+							},
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway 1",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway 2",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener 1",
+								Routes: []structs.ResourceReference{},
+							},
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway 2",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway 1",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener 2",
+						},
+						{
+							Name:        "Gateway 2",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener 2",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway 1",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener 1",
+							Routes: []structs.ResourceReference{},
+						},
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+				{
+					Name: "Gateway 2",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener 1",
+							Routes: []structs.ResourceReference{},
+						},
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route swaps from one listener to another on a gateway": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name: "Listener 1",
+								Routes: []structs.ResourceReference{
+									{
+										Name:        "TCP Route",
+										Kind:        structs.TCPRoute,
+										SectionName: "",
+									},
+								},
+							},
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener 2",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener 1",
+							Routes: []structs.ResourceReference{},
+						},
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"Multiple TCP Routes bind to different gateways": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway 1",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener 1",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway 1",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway 2",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway 2",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route 1",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway 1",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener 1",
+						},
+					},
+				},
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route 2",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway 2",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener 2",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+					Name: "Gateway 1",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener 1",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route 1",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+				{
+					Name: "Gateway 2",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route 2",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route cannot be bound to a listener with an HTTP protocol": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener",
+								Protocol: structs.ListenerProtocolHTTP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "Listener",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{},
+			expectedReferenceErrors: map[structs.ResourceReference]error{
+				{
+					Name:        "Gateway",
+					Kind:        structs.APIGateway,
+					SectionName: "Listener",
+				}: fmt.Errorf("failed to bind route TCP Route to gateway Gateway: listener Listener is not a tcp listener"),
+			},
+		},
+		"If a route/listener protocol mismatch occurs with the wildcard, but a bind to another listener was possible, no error is returned": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener 1",
+								Routes: []structs.ResourceReference{},
+							},
+							{
+								Name:   "Listener 2",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener 1",
+								Protocol: structs.ListenerProtocolHTTP,
+							},
+							{
+								Name:     "Listener 2",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{
+				{
+
+					Name: "Gateway",
+					Listeners: []structs.BoundAPIGatewayListener{
+						{
+							Name:   "Listener 1",
+							Routes: []structs.ResourceReference{},
+						},
+						{
+							Name: "Listener 2",
+							Routes: []structs.ResourceReference{
+								{
+									Name:        "TCP Route",
+									Kind:        structs.TCPRoute,
+									SectionName: "",
+								},
+							},
+						},
+					},
+				},
+			},
+			expectedReferenceErrors: map[structs.ResourceReference]error{},
+		},
+		"TCP Route references a listener that does not exist": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name:   "Listener",
+								Routes: []structs.ResourceReference{},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name:        "Gateway",
+							Kind:        structs.APIGateway,
+							SectionName: "Non-existent Listener",
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{},
+			expectedReferenceErrors: map[structs.ResourceReference]error{
+				{
+					Name:        "Gateway",
+					Kind:        structs.APIGateway,
+					SectionName: "Non-existent Listener",
+				}: fmt.Errorf("failed to bind route TCP Route to gateway Gateway with listener 'Non-existent Listener'"),
+			},
+		},
+		"Already bound TCP Route": {
+			gateways: []*gatewayMeta{
+				{
+					BoundGateway: &structs.BoundAPIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.BoundAPIGatewayListener{
+							{
+								Name: "Listener",
+								Routes: []structs.ResourceReference{{
+									Kind: structs.TCPRoute,
+									Name: "TCP Route",
+								}},
+							},
+						},
+					},
+					Gateway: &structs.APIGatewayConfigEntry{
+						Name: "Gateway",
+						Listeners: []structs.APIGatewayListener{
+							{
+								Name:     "Listener",
+								Protocol: structs.ListenerProtocolTCP,
+							},
+						},
+					},
+				},
+			},
+			routes: []structs.BoundRoute{
+				&structs.TCPRouteConfigEntry{
+					Name: "TCP Route",
+					Kind: structs.TCPRoute,
+					Parents: []structs.ResourceReference{
+						{
+							Name: "Gateway",
+							Kind: structs.APIGateway,
+						},
+					},
+				},
+			},
+			expectedBoundAPIGateways: []*structs.BoundAPIGatewayConfigEntry{},
+			expectedReferenceErrors:  map[structs.ResourceReference]error{},
+		},
+	}
+
+	for name, tc := range cases {
+		t.Run(name, func(t *testing.T) {
+			for i := range tc.gateways {
+				tc.gateways[i].initialize()
+			}
+
+			actualBoundAPIGateways, _, referenceErrors := bindRoutesToGateways(tc.gateways, tc.routes...)
+
+			require.Equal(t, tc.expectedBoundAPIGateways, actualBoundAPIGateways)
+			require.Equal(t, tc.expectedReferenceErrors, referenceErrors)
+		})
+	}
+}
+
 func TestAPIGatewayController(t *testing.T) {
 	conditions := newGatewayConditionGenerator()
 	defaultMeta := acl.DefaultEnterpriseMeta()
